Add the wind direction
Make the display for the wind speed round to make it easier to read, and
add in the wind direction.

This is kinda tricky because wind speed and direction are two different
messages on the MQTT broker. (That's just how the weather station
delivers it, it's not something we're doing.)

To do this, we're keeping the current wind speed and direction on the
stack and updating the entire message each time one or the other
changes.